fix: component naming. get rid of positioning inside components names
